meeting_assistant_mcp/requirements.txt
fastmcp==0.2.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2023.3
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
typing-extensions==4.8.0
